llm-phoenix-app/pyproject.toml
[project]
name = "llm-phoenix-app"
version = "0.1.0"
description = ""
authors = [
    {name = "kumar nitesh",email = "niteshgnit3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.14"
dependencies = [
    "arize-phoenix (>=10.0.4,<11.0.0)",
    "openai (>=1.82.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openinference-instrumentation-openai (>=0.1.28,<0.2.0)",
    "qdrant-client (>=1.14.2,<2.0.0)",
    "opentelemetry-sdk (>=1.33.1,<2.0.0)",
    "opentelemetry-exporter-otlp-proto-http (>=1.33.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
